add gcc support

+++ src/bigint/bigint.cpp
@@ -9,7 +9,7 @@ BigInt::BigInt(const std::string& number) : is_negative{} {
     this->set(std::string{number});
 }
 
-BigInt::BigInt(const long long number) : BigInt(std::to_string(number)) { }
+BigInt::BigInt(const int64_t number) : BigInt(std::to_string(number)) { }
 
 std::string BigInt::to_string() const {
     if (is_negative) {
